Fixed some error and issues

+++ FoundationKit/FoundationKit/Base/DateTime.h
@@ -621,7 +621,7 @@ public:
 	 * @param dateTime The date and time to get the hash for.
 	 * @return Hash value.
 	 */
-    uint32 getHash()
+    size_t getHash()
 	{
         return std::hash<int64>()(this->_ticks);
 	}
